\name{ReGenesees.GUI}
\alias{ReGenesees.GUI}
\title{Tcl/Tk GUI of the ReGenesees System}
\description{
Starts a graphical user interface that allows users to access the functionalities of the \pkg{ReGenesees} package.
}
\usage{
ReGenesees.GUI()
}

\details{
The main window of the GUI consists of a menu bar, a scrollable text window called "Commands Window" and a lower scrollable text window called "Warnings Window". The menu bar consists, from left to right, of the following menu items: "File", "Data", "Functions", "Tools", "Options", "Help" and "Website" (see section \sQuote{Menu Tree}).\cr

Menus "Data", "Functions" and "Options" lead to dialog boxes that need further specifications, while the others ("File", "Tools", "Help" and "Website") directly trigger the corresponding command. All dialog boxes accessed from the "Data" and "Functions" menus are provided with a help button that links to appropriate help pages (whose content is drawn mostly from the \pkg{ReGenesees} package documentation). For quicker guidance, all the entries of dialog boxes accessed from the "Functions" menu have been supplied with tooltips: just hover the mouse pointer over \sQuote{circled question mark} icons.\cr

Output objects created by the GUI are stored in the user's workspace (i.e. the \code{.GlobalEnv}). Upon creation, each output object (or a corresponding summary description, if the object is large) is also printed at the \R console.\cr

Whenever a GUI dialog causes a new object to be generated in the user's workspace (either by importing data or by executing a statistical method), the "Functions" menu-tree gets automatically updated. Menu entries will be enabled or disabled as objects are created or removed via the GUI, depending on the class of these objects. Note that, on the contrary, this will not happen if objects are created or removed by statements typed at the \R command line. In such cases, users can update the "Functions" menu-tree by using its first entry \emph{"Update Active Functions"}.\cr

Whenever a command generated by a GUI dialog happens to (i) create/destroy an object, (ii) perform I/O operations, or (iii) change system/environment settings, the command itself is inserted and stored in the "Commands Window". For instance, this happens to every command triggered by dialogs from the "Functions" menu. Thus, the "Commands Window" acts as a persistent log of the statistical analyses carried out during the current \pkg{ReGenesees} session. The content of the "Commands Window" can be saved via \emph{File -> Save Commands}. Groups of statements displayed in the "Commands Window" can also be copied by using the \emph{Control-C} command or right-clicking the mouse.\cr

All the commands inserted and stored in the "Commands Window" are \emph{also} flushed to the \R history file (see e.g. \code{?history}) immediately upon execution. Therefore, users can freely \emph{alternate} commands typed directly at the \R command-line interface \emph{and} commands triggered by the \pkg{ReGenesees} GUI, and still get a comprehensive log of \emph{all} the operations performed by simply saving the \R history file (e.g., in \code{Rgui}, through \emph{File -> Save History...}). Note that this feature is currently limited to Windows platforms.\cr

Warning messages arising from any menu item are displayed in the "Warnings Window". Such messages are not persistently saved: generating a new set of warnings erases any information about the previous ones. Error messages, causing the current computation to be stopped, are always displayed by means of pop-up dialogs.\cr

It is possible to end a GUI session by either using the \emph{File -> Exit} menu or closing the \pkg{ReGenesees} GUI main window directly.\cr
}

\section{Menu Tree}{
The scheme reported below represents the \pkg{ReGenesees} system menu-tree (with \sQuote{u.c.} meaning that the corresponding dialog window or item is still \sQuote{under construction}).\cr
\cr
\tabular{lll}{
\tab\cr
| File \tab  \tab  \cr
| --------------------------------------\tab  \tab  \cr
  \tab | - Change Directory \tab  \cr
  \tab | - Load Workspace \tab  \cr
  \tab | - Save Workspace \tab  \cr
  \tab | - Save Commands \tab  \cr
  \tab | - Exit \tab  \cr
\tab\cr
| Data \tab  \tab  \cr
| --------------------------------------\tab  \tab  \cr
  \tab | - Import Datasets \tab  \cr
\tab | --------------------------------------\tab  \cr
  \tab  \tab | - From Text Files\cr
  \tab  \tab | - From Excel or Access Files\cr
  \tab | - Export Datasets \tab  \cr
\tab\cr
| Functions \tab  \tab  \cr
| --------------------------------------\tab  \tab  \cr
  \tab | - Update Active Functions \tab  \cr
  \tab | - Survey Design \tab  \cr
\tab | --------------------------------------\tab  \cr
  \tab  \tab | - Specify a Survey Design\cr
  \tab  \tab | - Collapse Strata to Eliminate Lonely PSUs\cr
  \tab  \tab | - Add New Variables to a Survey Design\cr
  \tab  \tab | - Merge New Data into a Survey Design\cr
  \tab | - Calibration \tab  \cr
\tab | --------------------------------------\tab  \cr
  \tab  \tab | - Build a Template for Population Totals\cr
  \tab  \tab | - Describe a Population Totals Template\cr
  \tab  \tab | - Fill a Population Totals Template\cr
  \tab  \tab | - Calibrate a Survey Design\cr
  \tab  \tab | - Check for Calibration Convergence\cr
  \tab  \tab | - Trim Calibration Weights\cr
  \tab  \tab | - Calibration Models Options\cr
  \tab | - Estimates and Errors \tab  \cr
\tab | --------------------------------------\tab  \cr
  \tab  \tab | - Totals or Means\cr
  \tab  \tab | - Ratios\cr
  \tab  \tab | - Shares\cr
  \tab  \tab | - Share Ratios\cr
  \tab  \tab | - Regression Coefficients\cr
  \tab  \tab | - Quantiles\cr
  \tab  \tab | - Complex Estimators\cr
  \tab  \tab | - Auxiliary Variables Totals\cr
  \tab  \tab | - Variance Estimation Options\cr
\tab\cr
\tab\cr
\tab\cr
\tab\cr
\tab\cr
\tab\cr
\tab\cr
| Tools \tab  \tab  \cr
| --------------------------------------\tab  \tab  \cr
  \tab | - Browse Workspace \tab  \cr
  \tab | - Show Datasets \tab  \cr
  \tab | - Remove Objects \tab  \cr
\tab\cr
| Options \tab  \tab  \cr
| --------------------------------------\tab  \tab  \cr
  \tab | - Environment Options \tab  \cr
\tab\cr
| Help \tab  \tab  \cr
| --------------------------------------\tab  \tab  \cr
  \tab | - ReGenesees Reference Manual \tab  \cr
  \tab | - ReGenesees Flyer\tab  \cr
  \tab | - About the GUI \tab  \cr
\tab\cr
| Website \tab  \tab  \cr
| --------------------------------------\tab  \tab  \cr
  \tab | - ReGenesees website on GITHUB \tab  \cr
\tab\cr
    }
}

\author{
Diego Zardetto and Raffaella Cianchetta
}

\note{
The GUI requires a screen resolution of at least 1024x768: this may rule out some netbooks.\cr
The GUI is known to work better under the SDI (single document interface) mode of Rgui.\cr 
The GUI has been extensively tested under Windows OS. Tests on Mac OS X revealed that all GUI functionalities work correctly, although its overall appearance is noticeably worse than under Windows.
}

\examples{
# Start ReGenesees GUI
\dontrun{
ReGenesees.GUI()
}
}